revert changes to VUDefaultProvider; add checkFinite to VectorUtil instead, and call from KFVF.createType
